middle::entry -> rustc_passes
